backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
motor==3.3.1
python-multipart>=0.0.9
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.0
fpdf2>=2.7.0
openai>=1.0.0
httpx>=0.25.0
aiofiles>=23.0.0
python-jose>=3.3.0
resend>=0.5.0
dnspython>=2.4.0
